Split teams types
