+++ packages/types/src/index.ts
@@ -24,6 +24,7 @@ export * from './discord/stickers.js'
 export * from './discord/subscriptions.js'
 export * from './discord/users.js'
 export * from './discord/voice.js'
+export * from './discord/teams.js'
 export * from './discord/webhookEvents.js'
 export * from './discord/webhooks.js'
 export * from './discordeno.js'
